#527 Extract heatmap and overview components, style controls
