+++ angular-client/src/pages/bms-debug-page/components/hex-tile/hex-tile.component.html
@@ -1,4 +1,4 @@
-<div class="hex-tile" [style.background-color]="color()" [class.selected]="boxShadowColor()">
+<div [class]="viewClass()" [style.background-color]="color()" [class.selected]="boxShadowColor()">
   <span class="cell-number">{{ cellNumber() }}</span>
   <span class="cell-value-container">
     <span class="cell-value">{{ displayValue() }}</span>
