+++ angular-client/src/pages/bms-debug-page/components/segment-heatmap/segment-heatmap.component.html
@@ -1,0 +1,30 @@
+<div class="hex-grid-zone">
+  <div class="hex-row top-row">
+    @for (cell of topRowCells; track $index) {
+      <hex-tile
+        variant="heatmap-cell"
+        [value]="cell.value"
+        [booleanValue]="cell.boolValue"
+        [color]="getColor(cell)"
+        [currentView]="view"
+        [boxShadowColor]="isSelected(cell.reading)"
+        [cellNumber]="cell.cellNum"
+        (click)="cellClicked(cell.reading)"
+      />
+    }
+  </div>
+  <div class="hex-row bottom-row">
+    @for (cell of bottomRowCells; track $index) {
+      <hex-tile
+        variant="heatmap-cell"
+        [value]="cell.value"
+        [booleanValue]="cell.boolValue"
+        [color]="getColor(cell)"
+        [currentView]="view"
+        [boxShadowColor]="isSelected(cell.reading)"
+        [cellNumber]="cell.cellNum"
+        (click)="cellClicked(cell.reading)"
+      />
+    }
+  </div>
+</div>
