+++ angular-client/src/pages/bms-debug-page/components/segment-overview/segment-overview.component.html
@@ -1,0 +1,14 @@
+<div class="overview-zone">
+  <div class="stat">
+    <span class="stat-value">{{ formatTemp() }}<sup class="stat-unit">°C</sup></span>
+    <span class="stat-label">Temperature</span>
+  </div>
+  <div class="stat">
+    <span class="stat-value">{{ formatVoltage() }}<sup class="stat-unit">V</sup></span>
+    <span class="stat-label">Voltage</span>
+  </div>
+  <div class="stat">
+    <span class="stat-value">{{ formatChipTemp() }}<sup class="stat-unit">°C</sup></span>
+    <span class="stat-label">Chip Temp</span>
+  </div>
+</div>
